# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DevPilot API
  version: v1
servers:
- url: /
security:
- bearerAuth: []
paths:
...
  /api/v1/onboarding:
    post:
      operationId: onboardingComplete
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingResponse"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - onboarding
  /api/v1/plans:
    get:
      operationId: planList
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: cursor
        required: false
        schema:
          type: string
          maxLength: 512
          minLength: 0
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPagePlanSummaryView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - plan
  /api/v1/plans/active:
    get:
      operationId: planGetActive
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - plan
  /api/v1/plans/{planId}:
    get:
      operationId: planGet
      parameters:
      - in: path
        name: planId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - plan
  /api/v1/plans/{planId}/milestones/{milestoneId}:
    patch:
      operationId: planUpdateMilestone
      parameters:
      - in: path
        name: planId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: milestoneId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestonePatchRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - plan
  /api/v1/plans/{planId}/replan:
    post:
      operationId: planReplan
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      - in: path
        name: planId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplanRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplanCommitResponse"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - plan
  /api/v1/side-projects:
    get:
      operationId: projectList
      parameters:
      - in: query
        name: status
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - PAUSED
          - DONE
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - in: query
        name: cursor
        required: false
        schema:
          type: string
          maxLength: 512
          minLength: 0
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPageSideProjectView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - project
    post:
      operationId: projectCreate
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SideProjectCreateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SideProjectView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - project
  /api/v1/side-projects/{sideProjectId}:
    delete:
      operationId: projectDelete
      parameters:
      - in: path
        name: sideProjectId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - project
    get:
      operationId: projectGet
      parameters:
      - in: path
        name: sideProjectId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SideProjectView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - project
    patch:
      operationId: projectUpdate
      parameters:
      - in: path
        name: sideProjectId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SideProjectPatchRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SideProjectView"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Problem Details (docs/05 §1.2)
      tags:
      - project
...
components:
  schemas:
...
    CursorPagePlanSummaryView:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlanSummaryView"
        nextCursor:
          type: string
      required:
      - items
      - nextCursor
    CursorPageSideProjectView:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SideProjectView"
        nextCursor:
          type: string
      required:
      - items
      - nextCursor
    DiagnosticSuggestionView:
      type: object
      properties:
        category:
          type: string
          enum:
          - JAVA
          - SPRING
          - DATABASE
          - WEB_HTTP
          - NETWORK
          - CS
          - ALGORITHM
          - TESTING
          - DEVOPS
          - SECURITY
          - PRACTICAL_ENGINEERING
          - SYSTEM_DESIGN
          - EXPLANATION
        challengeId:
          type: string
          format: uuid
        difficulty:
          type: integer
          format: int32
        estimatedMinutes:
          type: integer
          format: int32
        selfAssessedLevel:
          type: integer
          format: int32
        skill:
          $ref: "#/components/schemas/SkillRef"
        title:
          type: string
      required:
      - category
      - challengeId
      - difficulty
      - estimatedMinutes
      - selfAssessedLevel
      - skill
      - title
    FieldError:
      type: object
      properties:
        code:
          type: string
        field:
          type: string
        message:
          type: string
      required:
      - code
      - field
      - message
    LearningGoalInput:
      type: object
      properties:
        checkpointDate:
          type: string
          format: date
        focusSkillCodes:
          type: array
          items:
            type: string
            maxLength: 100
            minLength: 0
          maxItems: 10
          minItems: 0
        targetCompletionDate:
          type: string
          format: date
        targetRole:
          type: string
          enum:
          - JAVA_BACKEND
      required:
      - focusSkillCodes
      - targetCompletionDate
      - targetRole
...
    LearningGoalView:
      type: object
      properties:
        checkpointDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        focusSkills:
          type: array
          items:
            $ref: "#/components/schemas/SkillRef"
        id:
          type: string
          format: uuid
        replanRecommended:
          type: boolean
        targetCompletionDate:
          type: string
          format: date
        targetRole:
          type: string
          enum:
          - JAVA_BACKEND
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
      required:
      - checkpointDate
      - createdAt
      - focusSkills
      - id
      - replanRecommended
      - targetCompletionDate
      - targetRole
      - updatedAt
      - version
...
    MilestoneIdMappingView:
      type: object
      properties:
        newId:
          type: string
          format: uuid
        previousId:
          type: string
          format: uuid
      required:
      - newId
      - previousId
...
    MilestonePatchRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 2000
          minLength: 0
        sortOrder:
          type: integer
          format: int32
          maximum: 10000
          minimum: 0
        status:
          type: string
          enum:
          - PLANNED
          - IN_PROGRESS
          - DONE
          - DEFERRED
          - DROPPED
        version:
          type: integer
          format: int64
      required:
      - version
    MilestoneView:
      type: object
      properties:
        description:
          type: string
        endDate:
          type: string
          format: date
        id:
          type: string
          format: uuid
        priority:
          type: string
          enum:
          - MUST
          - SHOULD
          - LATER
        skillCodes:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
          format: int32
        startDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - PLANNED
          - IN_PROGRESS
          - DONE
          - DEFERRED
          - DROPPED
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
      required:
      - description
      - endDate
      - id
      - priority
      - skillCodes
      - sortOrder
      - startDate
      - status
      - title
      - updatedAt
      - version
    OnboardingRequest:
      type: object
      properties:
        dayStartHour:
          type: integer
          format: int32
          maximum: 6
          minimum: 0
        displayName:
          type: string
          maxLength: 100
          minLength: 0
        experienceProfile:
          type: string
          enum:
          - WORKING_DEVELOPER
          - DEVELOPER_STARTER
          - OTHER
        experienceStartDate:
          type: string
          format: date
        learningGoal:
          $ref: "#/components/schemas/LearningGoalInput"
        runDiagnostic:
          type: boolean
        selfAssessments:
          type: array
          items:
            $ref: "#/components/schemas/SelfAssessmentInput"
          maxItems: 13
          minItems: 0
        sideProject:
          $ref: "#/components/schemas/SideProjectInput"
        timezone:
          type: string
          maxLength: 50
          minLength: 0
        useTemplate:
          type: boolean
        weekdayStudyMinutes:
          type: integer
          format: int32
          maximum: 720
          minimum: 0
        weekendStudyMinutes:
          type: integer
          format: int32
          maximum: 720
          minimum: 0
      required:
      - dayStartHour
      - displayName
      - experienceProfile
      - learningGoal
      - runDiagnostic
      - selfAssessments
      - timezone
      - useTemplate
      - weekdayStudyMinutes
      - weekendStudyMinutes
    OnboardingResponse:
      type: object
      properties:
        activePlan:
          $ref: "#/components/schemas/PlanSummaryView"
        assignedSeedCardCount:
          type: integer
          format: int32
        learningGoal:
          $ref: "#/components/schemas/LearningGoalView"
        sideProject:
          $ref: "#/components/schemas/SideProjectView"
        suggestedDiagnostics:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticSuggestionView"
        user:
          $ref: "#/components/schemas/MeResponse"
      required:
      - activePlan
      - assignedSeedCardCount
      - learningGoal
      - sideProject
      - suggestedDiagnostics
      - user
...
    PlanSummaryView:
      type: object
      properties:
        changeReason:
          type: string
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        latestRatioBp:
          type: integer
          format: int32
        latestRiskLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
        milestoneCount:
          type: integer
          format: int32
        planVersion:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - ACTIVE
          - SUPERSEDED
          - ARCHIVED
        supersededAt:
          type: string
          format: date-time
        title:
          type: string
      required:
      - changeReason
      - createdAt
      - id
      - latestRatioBp
      - latestRiskLevel
      - milestoneCount
      - planVersion
      - status
      - supersededAt
      - title
    PlanView:
      type: object
      properties:
        changeReason:
          type: string
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        latestSnapshot:
          $ref: "#/components/schemas/SnapshotView"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/MilestoneView"
        planVersion:
          type: integer
          format: int32
        replanRecommended:
          type: boolean
        skillTargets:
          type: array
          items:
            $ref: "#/components/schemas/PlanSkillTargetView"
        status:
          type: string
          enum:
          - ACTIVE
          - SUPERSEDED
          - ARCHIVED
        supersededAt:
          type: string
          format: date-time
        supersedesPlanId:
          type: string
          format: uuid
        title:
          type: string
        version:
          type: integer
          format: int64
      required:
      - changeReason
      - createdAt
      - id
      - latestSnapshot
      - milestones
      - planVersion
      - replanRecommended
      - skillTargets
      - status
      - supersededAt
      - supersedesPlanId
      - title
      - version
    ProblemDetail:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        instance:
          type: string
        status:
          type: integer
          format: int32
        title:
          type: string
        traceId:
          type: string
        type:
          type: string
      required:
      - code
      - detail
      - errors
      - instance
      - status
      - title
      - traceId
      - type
    ReplanCommitResponse:
      type: object
      properties:
        milestoneIdMapping:
          type: array
          items:
            $ref: "#/components/schemas/MilestoneIdMappingView"
        plan:
          $ref: "#/components/schemas/PlanView"
      required:
      - milestoneIdMapping
      - plan
    ReplanRequest:
      type: object
      properties:
        acceptedDeferrals:
          type: array
          items:
            type: string
            maxLength: 100
            minLength: 0
          maxItems: 100
          minItems: 0
        acceptedTargetRaises:
          type: array
          items:
            $ref: "#/components/schemas/TargetRaiseInput"
          maxItems: 400
          minItems: 0
        acceptedTargetReductions:
          type: array
          items:
            $ref: "#/components/schemas/TargetReductionInput"
          maxItems: 400
          minItems: 0
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/MilestoneInput"
          maxItems: 24
          minItems: 0
        reason:
          type: string
          maxLength: 1000
          minLength: 0
        restoredDeferrals:
          type: array
          items:
            type: string
            maxLength: 100
            minLength: 0
          maxItems: 100
          minItems: 0
        version:
          type: integer
          format: int64
      required:
      - acceptedDeferrals
      - acceptedTargetRaises
      - acceptedTargetReductions
      - milestones
      - reason
      - restoredDeferrals
      - version
...
    SelfAssessmentInput:
      type: object
      properties:
        category:
          type: string
          enum:
          - JAVA
          - SPRING
          - DATABASE
          - WEB_HTTP
          - NETWORK
          - CS
          - ALGORITHM
          - TESTING
          - DEVOPS
          - SECURITY
          - PRACTICAL_ENGINEERING
          - SYSTEM_DESIGN
          - EXPLANATION
        level:
          type: integer
          format: int32
          maximum: 5
          minimum: 0
      required:
      - category
      - level
    SideProjectCreateRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 1000
          minLength: 0
        name:
          type: string
          maxLength: 100
          minLength: 0
        repoUrl:
          type: string
          maxLength: 500
          minLength: 0
        stack:
          type: string
          maxLength: 300
          minLength: 0
      required:
      - name
    SideProjectInput:
      type: object
      properties:
        description:
          type: string
          maxLength: 1000
          minLength: 0
        name:
          type: string
          maxLength: 100
          minLength: 0
        repoUrl:
          type: string
          maxLength: 500
          minLength: 0
        stack:
          type: string
          maxLength: 300
          minLength: 0
      required:
      - name
    SideProjectPatchRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 1000
          minLength: 0
        name:
          type: string
          maxLength: 100
          minLength: 0
        repoUrl:
          type: string
          maxLength: 500
          minLength: 0
        stack:
          type: string
          maxLength: 300
          minLength: 0
        status:
          type: string
          enum:
          - ACTIVE
          - PAUSED
          - DONE
        version:
          type: integer
          format: int64
      required:
      - version
    SideProjectView:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        description:
          type: string
        id:
          type: string
          format: uuid
        name:
          type: string
        repoUrl:
          type: string
        stack:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - PAUSED
          - DONE
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
      required:
      - createdAt
      - description
      - id
      - name
      - repoUrl
      - stack
      - status
      - updatedAt
      - version
...
    SnapshotView:
      type: object
      properties:
        completionRateBp:
          type: integer
          format: int32
        effectiveBudgetMinutes:
          type: integer
          format: int32
        generatedAt:
          type: string
          format: date-time
        horizonDate:
          type: string
          format: date
        nominalBudgetMinutes:
          type: integer
          format: int32
        ratioBp:
          type: integer
          format: int32
        requiredMustMinutes:
          type: integer
          format: int32
        requiredShouldMinutes:
          type: integer
          format: int32
        riskLevel:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
        snapshotDate:
          type: string
          format: date
      required:
      - completionRateBp
      - effectiveBudgetMinutes
      - generatedAt
      - horizonDate
      - nominalBudgetMinutes
      - ratioBp
      - requiredMustMinutes
      - requiredShouldMinutes
      - riskLevel
      - snapshotDate
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http